move schema to module
