+++ schema-dynamic.mjs
